requirements.txt
black==19.10b0
eth-brownie>=1.11.10,<2.0.0
sgqlc==11.0
dotmap==1.3.23
tabulate==0.8.7
ens-namehash==1.0.0
rich==9.3.0
boto3==1.16.28
python-dotenv==0.15.0
multicall==0.1.1